Feature - Car removed from game on when crossing specified waypoints
I have implemented the first half of the feature spawning and de-spawning the vehicles.

I have added a field into the waypoint script that is a boolean "carRemover" to say if the waypoint is a de-spawn point for cars. Then in the mobile object script when we switch waypoint if have added a code block that checks to see if the waypoint is a remover. If it is a remover then we call a method in the traffic controller and remove the car from the vehicles array and then we remove the gameObject using the unity Destroy() function.