in game main menu functionallity
I have completed the in game main menu functionallity. To access the in game menu press escape in the in game menu you can resume the game by either clicking the resume button or pressing escape again. In the in game menu you can access the options page to then adjust the in game volume levels or exit back to the main menu by clicking return to main menu.